requirements.txt
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.18.5
opencv-python>=4.5.1
Pillow>=8.0.0
scikit-learn>=0.24.0
